Tests/LibWeb/Text/input/hit_testing/open-details-by-clicking-on-triangle.html

<style>
    body {
        margin: 0;
    }
</style>
<details id="details"><summary><span>here be bugs</span></summary><span>you can't see me</span></details>
<script src="../include.js"></script>
<script>
    asyncTest(done => {
        details.ontoggle = (event) => {
            println(`toggle ${event.newState}`);
            done();
        };
        internals.click(5, 5);
    });
</script>
