Tests/LibWeb/Text/expected/hit_testing/open-details-by-clicking-on-triangle.txt

here be bugsyou can't see me   toggle open
